dialog-host: Fix allocating children
Clamping child size to the minimum requisition as returned by
gtk_widget_get_preferred_size results in over-allocating the child even
when the original size was perfectly acceptable to begin with. Instead
of doing that, use the newly added adw_ensure_child_allocation_size to
only extend the allocation when the original size is not enough to fit
it.

Also, since we're only measuring the primary child in our measure (),
propagate its request mode, instead of looking for a common request mode
among our children.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>