dialog: always clear priv->last_focus weak pointer
Currently we're relying on the last-focused widget to never outlive the
AdwDialog. In fairness, this would be unusual, but it's always possible
and the penalty for being wrong is memory corruption, and that is a very
harsh penalty.


(cherry picked from commit 406c76d12d8b4737d4e41e007ed77aebb77a1ac7)

Co-authored-by: Michael Catanzaro <mcatanzaro@redhat.com>