stylesheet: Fix oversized GtkPaned handle drag area
The drag catch area of GtkPaned is supposed to be 6px on
each side of the handle, but the CSS adds an extra 8px on the
end-child side.

This seems to be a leftover from gtk3.

Remove the margins and padding, to avoid preempting pointer clicks
on a too wide area of the end-child.

Relates gtk#4484