toast-overlay: fix assertion failure when showing toast while hiding it
This is a weird scenario, but if we show a toast while its hide
animation is still playing, then we crash.

I've failed to reproduce this scenario and I've also failed to quickly
write an API test, though it really ought to be possible, so this is a
speculative fix. But it looks right.

Probably fixes #997